bruig: improve chat scroll controller + code reorg
