+++ bruig/flutterui/bruig/lib/components/chat/events.dart
@@ -1,308 +1,25 @@
 import 'dart:async';
 
-import 'package:bruig/components/attach_file.dart';
+import 'package:flutter/material.dart';
+import 'package:bruig/models/client.dart';
+import 'package:bruig/components/chat/types.dart';
+import 'package:golib_plugin/definitions.dart';
 import 'package:bruig/components/buttons.dart';
-import 'package:bruig/components/chats_list.dart';
-import 'package:bruig/components/manage_gc.dart';
-import 'package:bruig/components/snackbars.dart';
-import 'package:bruig/screens/feed/feed_posts.dart';
 import 'package:bruig/components/md_elements.dart';
 import 'package:bruig/components/user_content_list.dart';
-import 'package:bruig/models/client.dart';
 import 'package:bruig/models/downloads.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:golib_plugin/util.dart';
 import 'package:intl/intl.dart';
-import 'package:golib_plugin/definitions.dart';
 import 'package:golib_plugin/golib_plugin.dart';
 import 'package:provider/provider.dart';
-import 'package:bruig/components/profile.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:bruig/components/empty_widget.dart';
 import 'package:open_filex/open_filex.dart';
 import 'package:file_icon/file_icon.dart';
 import 'package:bruig/components/interactive_avatar.dart';
 import 'package:bruig/components/user_context_menu.dart';
-
-class ActiveChat extends StatefulWidget {
-  final ClientModel client;
-  final FocusNode editLineFocusNode;
-  ActiveChat(this.client, this.editLineFocusNode, {Key? key}) : super(key: key);
-
-  @override
-  State<ActiveChat> createState() => _ActiveChatState();
-}
-
-class _ActiveChatState extends State<ActiveChat> {
-  ClientModel get client => widget.client;
-  FocusNode get editLineFocusNode => widget.editLineFocusNode;
-  ChatModel? chat;
-
-  void clientChanged() {
-    var newChat = client.active;
-    if (newChat != chat) {
-      setState(() {
-        chat = newChat;
-      });
-    }
-  }
-
-  void sendMsg(String msg) async {
-    try {
-      await chat?.sendMsg(msg);
-    } catch (exception) {
-      if (mounted) {
-        showErrorSnackbar(context, "Unable to send message: $exception");
-      }
-    }
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    chat = client.active;
-    client.addListener(clientChanged);
-  }
-
-  @override
-  void didUpdateWidget(ActiveChat oldWidget) {
-    oldWidget.client.removeListener(clientChanged);
-    super.didUpdateWidget(oldWidget);
-    client.addListener(clientChanged);
-  }
-
-  @override
-  void dispose() {
-    client.removeListener(clientChanged);
-    editLineFocusNode.dispose();
-    super.dispose();
-  }
-
-  String nickCapitalLetter() =>
-      chat != null && chat!.nick.isNotEmpty ? chat!.nick[0].toUpperCase() : "";
-
-  @override
-  Widget build(BuildContext context) {
-    if (this.chat == null) return Container();
-    var chat = this.chat!;
-
-    var profile = client.profile;
-    if (profile != null) {
-      if (chat.isGC) {
-        return const ManageGCScreen();
-      } else {
-        return UserProfile(client, profile);
-      }
-    }
-    //editLineFocusNode.requestFocus();
-    var theme = Theme.of(context);
-    var textColor = theme.dividerColor;
-    var darkTextColor = theme.indicatorColor;
-    var selectedBackgroundColor = theme.highlightColor;
-    var subMenuBorderColor = theme.canvasColor;
-    var hightLightTextColor = theme.focusColor; // NAME TEXT COLOR
-    var avatarColor = colorFromNick(chat.nick);
-    var avatarTextColor =
-        ThemeData.estimateBrightnessForColor(avatarColor) == Brightness.dark
-            ? hightLightTextColor
-            : darkTextColor;
-
-    showSubMenu(String id) => chat.isGC
-        ? client.showSubMenu(true, id)
-        : client.showSubMenu(false, id);
-
-    return Row(children: [
-      Expanded(
-          child: Column(children: [
-        Expanded(child: Messages(chat, client.nick, showSubMenu, client)),
-        Row(children: [
-          Expanded(child: EditLine(sendMsg, chat, editLineFocusNode))
-        ])
-      ])),
-      client.activeSubMenu.isEmpty
-          ? const Empty()
-          : Container(
-              width: 250,
-              decoration: BoxDecoration(
-                  border: Border(
-                      left: BorderSide(width: 2, color: subMenuBorderColor))),
-              child: Stack(alignment: Alignment.topRight, children: [
-                Column(children: [
-                  Container(
-                      margin: const EdgeInsets.only(top: 20, bottom: 20),
-                      child: CircleAvatar(
-                          radius: 75,
-                          backgroundColor: avatarColor,
-                          child: Text(nickCapitalLetter(),
-                              style: TextStyle(
-                                  color: avatarTextColor, fontSize: 75)))),
-                  chat.isGC
-                      ? Text("Group Chat",
-                          style: TextStyle(fontSize: 15, color: textColor))
-                      : Empty(),
-                  Text(chat.nick,
-                      style: TextStyle(fontSize: 15, color: textColor)),
-                  ListView.builder(
-                    shrinkWrap: true,
-                    itemCount: client.activeSubMenu.length,
-                    itemBuilder: (context, index) => ListTile(
-                        title: Text(client.activeSubMenu[index].label,
-                            style: const TextStyle(fontSize: 11)),
-                        onTap: () {
-                          client.activeSubMenu[index]
-                              .onSelected(context, client);
-                          client.hideSubMenu();
-                        },
-                        hoverColor: Colors.black),
-                  )
-                ]),
-                Positioned(
-                    top: 5,
-                    right: 5,
-                    child: Material(
-                        color: selectedBackgroundColor.withOpacity(0),
-                        child: IconButton(
-                            tooltip: "Close",
-                            hoverColor: selectedBackgroundColor,
-                            splashRadius: 15,
-                            iconSize: 15,
-                            onPressed: () => client.hideSubMenu(),
-                            icon: Icon(
-                                color: darkTextColor, Icons.close_outlined)))),
-              ]))
-    ]);
-  }
-}
-
-typedef SendMsg = void Function(String msg);
-
-class EditLine extends StatefulWidget {
-  final SendMsg _send;
-  final ChatModel chat;
-  final FocusNode editLineFocusNode;
-  EditLine(this._send, this.chat, this.editLineFocusNode, {Key? key})
-      : super(key: key);
-
-  @override
-  State<EditLine> createState() => _EditLineState();
-}
-
-class _EditLineState extends State<EditLine> {
-  final controller = TextEditingController();
-
-  final FocusNode node = FocusNode();
-  List<AttachmentEmbed> embeds = [];
-
-  @override
-  void initState() {
-    super.initState();
-    controller.text = widget.chat.workingMsg;
-  }
-
-  @override
-  void didUpdateWidget(EditLine oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    var workingMsg = widget.chat.workingMsg;
-    if (workingMsg != controller.text) {
-      controller.text = workingMsg;
-      controller.selection = TextSelection(
-          baseOffset: workingMsg.length, extentOffset: workingMsg.length);
-      widget.editLineFocusNode.requestFocus();
-    }
-  }
-
-  void handleKeyPress(event) {
-    if (event is RawKeyUpEvent) {
-      bool modPressed = event.isShiftPressed || event.isControlPressed;
-      final val = controller.value;
-      if (event.data.logicalKey.keyLabel == "Enter" && !modPressed) {
-        final messageWithoutNewLine =
-            controller.text.substring(0, val.selection.start - 1) +
-                controller.text.substring(val.selection.start);
-        controller.value = const TextEditingValue(
-            text: "", selection: TextSelection.collapsed(offset: 0));
-        final String withEmbeds = embeds.fold(
-            messageWithoutNewLine.trim(), (s, e) => e.replaceInString(s));
-        /*
-          if (withEmbeds.length > 1024 * 1024) {
-            showErrorSnackbar(context,
-                "Message is larger than maximum allowed (limit: 1MiB)");
-            return;
-          }
-          */
-        if (withEmbeds != "") {
-          widget._send(withEmbeds);
-          widget.chat.workingMsg = "";
-          setState(() {
-            embeds = [];
-          });
-        }
-      } else {
-        widget.chat.workingMsg = val.text.trim();
-      }
-    }
-  }
-
-  void attachFile() async {
-    var res = await Navigator.of(context, rootNavigator: true)
-        .pushNamed(AttachFileScreen.routeName);
-    if (res == null) {
-      return;
-    }
-    var embed = res as AttachmentEmbed;
-    embeds.add(embed);
-    setState(() {
-      controller.text = controller.text + embed.displayString() + " ";
-      widget.chat.workingMsg = controller.text;
-      widget.editLineFocusNode.requestFocus();
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var theme = Theme.of(context);
-    var textColor = theme.focusColor; // MESSAGE TEXT COLOR
-    var hoverColor = theme.hoverColor;
-    var backgroundColor = theme.highlightColor;
-    var hintTextColor = theme.dividerColor;
-    return RawKeyboardListener(
-        focusNode: node,
-        onKey: handleKeyPress,
-        child: Container(
-            margin: const EdgeInsets.only(bottom: 5),
-            child: Row(children: [
-              IconButton(onPressed: attachFile, icon: Icon(Icons.attach_file)),
-              Expanded(
-                  child: TextField(
-                autofocus: true,
-                focusNode: widget.editLineFocusNode,
-                style: TextStyle(
-                  fontSize: 11,
-                  color: textColor,
-                ),
-                controller: controller,
-                minLines: 1,
-                maxLines: null,
-                //textInputAction: TextInputAction.done,
-                //style: normalTextStyle,
-                keyboardType: TextInputType.multiline,
-                decoration: InputDecoration(
-                  filled: true,
-                  fillColor: backgroundColor,
-                  hoverColor: hoverColor,
-                  isDense: true,
-                  hintText: 'Type a message',
-                  hintStyle: TextStyle(
-                    fontSize: 11,
-                    color: hintTextColor,
-                  ),
-                  border: InputBorder.none,
-                ),
-              )),
-            ])));
-  }
-}
+import 'package:bruig/components/empty_widget.dart';
+// TODO: isolate
+import 'package:bruig/screens/feed/feed_posts.dart';
 
 class ServerEvent extends StatelessWidget {
   final Widget child;
@@ -671,6 +388,7 @@ class _PostsListWState extends State<PostsListW> {
         Text("User Posts", style: TextStyle(color: textColor)),
         ListView.builder(
             shrinkWrap: true,
+            physics: const NeverScrollableScrollPhysics(),
             itemCount: posts.length,
             itemBuilder: (BuildContext context, int index) {
               return Row(
@@ -1149,13 +867,15 @@ class Event extends StatelessWidget {
   final ChatEventModel event;
   final ChatModel chat;
   final String nick;
-  final ShowSubMenuCB showSubMenu;
   final ClientModel client;
   final Function() scrollToBottom;
-  const Event(this.chat, this.event, this.nick, this.client,
-      this.scrollToBottom, this.showSubMenu,
+  const Event(
+      this.chat, this.event, this.nick, this.client, this.scrollToBottom,
       {Key? key})
       : super(key: key);
+
+  showSubMenu(String id) =>
+      chat.isGC ? client.showSubMenu(true, id) : client.showSubMenu(false, id);
 
   @override
   Widget build(BuildContext context) {
@@ -1237,92 +957,5 @@ class Event extends StatelessWidget {
         color: Theme.of(context).errorColor, // ERROR COLOR
         child: Text("Unknonwn chat event type",
             style: TextStyle(color: textColor)));
-  }
-}
-
-class Messages extends StatefulWidget {
-  final ChatModel chat;
-  final String nick;
-  final ShowSubMenuCB showSubMenu;
-  final ClientModel client;
-  const Messages(this.chat, this.nick, this.showSubMenu, this.client,
-      {Key? key})
-      : super(key: key);
-
-  @override
-  State<Messages> createState() => _MessagesState();
-}
-
-class _MessagesState extends State<Messages> {
-  ClientModel get client => widget.client;
-  ChatModel get chat => widget.chat;
-  String get nick => widget.nick;
-  final ScrollController _scroller = ScrollController();
-  bool _firstAutoscrollDone = false;
-  bool _shouldAutoscroll = false;
-
-  void scrollToBottom() {
-    _scroller.jumpTo(_scroller.position.maxScrollExtent);
-  }
-
-  void scrollListener() {
-    _firstAutoscrollDone = true;
-    if (_scroller.hasClients &&
-        _scroller.position.pixels == _scroller.position.maxScrollExtent) {
-      _shouldAutoscroll = true;
-    } else {
-      _shouldAutoscroll = false;
-    }
-  }
-
-  void maybeScrollToBottom() {
-    if (_scroller.hasClients && (_shouldAutoscroll || !_firstAutoscrollDone)) {
-      scrollToBottom();
-    }
-  }
-
-  void onChatChanged() {
-    setState(() {
-      maybeScrollToBottom();
-    });
-    Future.delayed(const Duration(milliseconds: 50), () {
-      setState(maybeScrollToBottom);
-    });
-  }
-
-  @override
-  initState() {
-    super.initState();
-    _scroller.addListener(scrollListener);
-    chat.addListener(onChatChanged);
-  }
-
-  @override
-  void didUpdateWidget(Messages oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    oldWidget.chat.removeListener(onChatChanged);
-    chat.addListener(onChatChanged);
-    _firstAutoscrollDone = false;
-    _shouldAutoscroll = true;
-    Future.delayed(const Duration(milliseconds: 1), () {
-      setState(maybeScrollToBottom);
-    });
-  }
-
-  @override
-  dispose() {
-    _scroller.removeListener(scrollListener);
-    chat.removeListener(onChatChanged);
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var msgs = chat.msgs; // Probably inneficient to regenerate every render...
-    return ListView.builder(
-        controller: _scroller,
-        itemCount: msgs.length,
-        itemBuilder: (context, index) => Event(chat, msgs[index], nick, client,
-            scrollToBottom, widget.showSubMenu));
   }
 }
