+++ bruig/flutterui/bruig/lib/components/chat/types.dart
@@ -1,0 +1,5 @@
+import 'package:bruig/models/client.dart';
+
+typedef SendMsg = void Function(String msg);
+typedef MakeActiveCB = void Function(ChatModel? c);
+typedef ShowSubMenuCB = void Function(String);
