+++ bruig/flutterui/bruig/lib/components/chats_list.dart
@@ -8,9 +8,7 @@ import 'package:bruig/components/snackbars.dart';
 import 'package:bruig/components/interactive_avatar.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:bruig/components/user_context_menu.dart';
-
-typedef MakeActiveCB = void Function(ChatModel? c);
-typedef ShowSubMenuCB = void Function(String);
+import 'package:bruig/components/chat/types.dart';
 
 class _ChatHeadingW extends StatefulWidget {
   final ChatModel chat;
