+++ bruig/flutterui/bruig/macos/Podfile
@@ -1,4 +1,4 @@
-platform :osx, '10.13'
+platform :osx, '10.14'
 
 # CocoaPods analytics sends network stats synchronously affecting flutter build latency.
 ENV['COCOAPODS_DISABLE_STATS'] = 'true'
