+++ bruig/flutterui/bruig/pubspec.yaml
@@ -49,6 +49,7 @@ dependencies:
   file_icon: ^1.0.0
   package_info_plus: ^3.0.3
   open_filex: ^4.3.2
+  scrollable_positioned_list: ^0.3.5
 
 msix_config:
   display_name: Bison Relay GUI
